Merged PR 1246: Fixing visualizations out of sync.
# Implementation Details
-------------------------------
## What was changed?
Visualizations in the statistics window are now in sync.

## Why was it changed?
Visualizations were out of sync because of the way threading works in python. We used to have a separate thread to process results. it was causing the issue.

## How was it changed?
Processing results is now done in the same thread as training. This will fix the problem.

# How to review
Create a model.
start training. check if visualizations are looking good

# Please review the following files
-------------------------------
all

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
removing extra thread for processing results will help us have more control over processing results or any other potential bugs in the future. It will reduce the uncertainties related to python threading since we don't have control over how the python switches between the threads.

removed unnecessary line in trainer.py

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack