Merged PR 1250: Set the datasetSettings to model
# Implementation Details
-------------------------------
## What was changed?
It was set dataset settings to model

## Why was it changed?
To have dataset settings whenever we want.

## How was it changed?
When model is created insted to set it to local storage it was set like a mode metadata.

# How to review
-------------------------------
Steps to reproduce:
close all the models in the workspace.
refresh the browser
create a new model with different dataset
now, go to modelhub and open some existing model which uses dataset different from the one in step 3. Or try to open a model whose previews aren’t stored in cache.

This will be applied to newly created models

# Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Removed some code which was used for electron version.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Set the datasetSettings to model

Related work items: #1963