Merged PR 1252: Start training in incognito
# Implementation Details
-------------------------------
## What was changed?
It was set dataset settings to model

## Why was it changed?
To have dataset settings whenever we want.

## How was it changed?
When model is created insted to set it to local storage it was set like a mode metadata.

# How to review
-------------------------------
Start the frontend in incognito/private mode
Create any model
Press Run
# Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
I was replaced in all places where is used data wizard env variable when is checked if it's true or false to a function from helpers file and cleand up some fn from select-model-modal

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #1960