Merged PR 1259: Rename Output -> Target
# Implementation Details
-------------------------------
## What was changed?
The 'output' datatype was renamed to 'target'

## Why was it changed?
To avoid confusion between outputs of a model and expected outputs (in the target/label sense)

## How was it changed?

We did the minimum work to fix it for the user.

Note that there are _a lot_ of references to e.g., IoOutput that was _not_ changed. The effort to rename everything at once is disproportinate to the benefits at this moment, so we'll introduce some debt and rename things as we touch upon them.

Thanks @<Anton Bourosu> for doing the frontend work.

# How to review
-------------------------------

Run the tool and check that the "output" layers now have the name "target"

# Please review the following files
-------------------------------
All of them.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
N/A.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #1786