Merged PR 1263: Questionnaire weren't close after confirm.
# Implementation Details
-------------------------------
## What was changed?
It was changed fn wich close the modal

## Why was it changed?
Because in samoe cases it doesn't close the quesstionnaire.

## How was it changed?
There was condition depend if was shown wast's new tips in in case that was false it wasn't close the quesstionnaire.

# How to review
-------------------------------
- Clear local storage
- Log out and create new account
- Confirm link from email and go trought the quesstionnaire

# Please review the following files
-------------------------------
*To be filled in*

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Was added commnet to fn in case we decide to enable back wat's new tips.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2018