Merged PR 1262: Navigate to new test route.
# Implementation Details
-------------------------------
## What was changed?
It was chenged the action that is happen when click run test in model hub.

## Why was it changed?
It was changed beacuse it was goes to old test view, which is no longer used.
## How was it changed?
It was changed to a link with test route.

# How to review
-------------------------------
Train a model, got to model hub and pres on run tests.

# Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
I had removed some part that was belong to old test view.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #1977