Merged PR 1260: Remove open ocde button for IO components
# Implementation Details
-------------------------------
## What was changed?
It was removed open code btn for IO components

## Why was it changed?
Beacause we don't need to have posibility to change anythin in IO components.
## How was it changed?
It was added a condition base on layer type of component.

# How to review
-------------------------------
Open any models and select a IO componet, in right sidebat suldn't be a open code btn.

# Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Was removed some console logs.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #1973