Merged PR 1274: Fix map component is spinning when run model directly
# Implementation Details
-------------------------------
## What was changed?
*To be filled in*

## Why was it changed?
*To be filled in*

## How was it changed?
*To be filled in*

# How to review
-------------------------------
1. Load any dataset
2. Press "Run Model"

For everything to look the same if you had:
Pressed "Customize"
Pressed "Run with Current Settings"

# Please review the following files
-------------------------------
*To be filled in*

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
There are many similar workflows and I think we can optimize them for sure.
For example to run a model, we have different workflows, but what they all do are really the similar,
We might need to think of a centralized workflow with various kinds of configuration.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Fix map component is spinning when run model directly

Related work items: #1955