Merged PR 1295: Implement static loading message and indicator for test
# Implementation Details
-------------------------------
## What was changed?
*To be filled in*

## Why was it changed?
*To be filled in*

## How was it changed?
*To be filled in*

# How to review
-------------------------------
1. Go to test view
2. Run tests
You should see a loading indicator and a static message while tests are running

# Please review the following files
-------------------------------
*To be filled in*

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Instead of creating an action for a mutation, name it with an action related to it, so that we can add new behaviors later.
e.g. "testStart" only for setting the running flag to true (but we will add new features later)

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Implement static loading message and indicator for test

Related work items: #2055