Merged PR 1265: Fix Export and Export to Github button enabled when no model is opened
# Implementation Details
-------------------------------
The request for this ticket is not meet the current design.
For now, if those actions work only for current model, not selected models in Model hub nor several models.
If we need to change, we may need a user story to change the workflow.

# How to review
-------------------------------
- Go to Model Hub
Export, Export To Github Actions in menu should be disabled

- Go to Modeling Tool(Or Statistics) and close all models opened
Export, Export To Github Actions in menu should be disabled

- On Modeling Tool (Or Statistics View), open a model.
Export, Export To Github Actions should work for current model.

# Please review the following files
-------------------------------
*To be filled in*

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
*To be filled in*

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Fix Export and Export to Github button enabled when no model is opened

Related work items: #1968