Merged PR 1283: Center select component, add label to it.
# Implementation Details
-------------------------------
## What was changed?
It was changed text align for select IO components and was added placeholder.

## Why was it changed?
To be more intuitively what user should do.

## How was it changed?
it was passed a class to select componet which center the items and passed a props for placeholder.

# How to review
-------------------------------
Load a csv.

# Please review the following files
-------------------------------
-

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
N/A

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Center select component, add label to it.

Related work items: #2047