Merged PR 1302: Kernel No Longer Hangs at 0.0001 Learning Rate
# Implementation Summary
-------------------------------
The changes in this branch seek to address the issue presented by this bug:

 #2069  It was caused by the learning rate being read into the kernel as a string rather than a float, and calling the optimizer with that incorrectly set parameter. To fix it, I added logic to check the type of the parameters passed in.

 How to review
-------------------------------
Create any arbitrary model with a learning rate of 0.001 in master, and you'll see a crash. Try it in this branch and it should work

Please review the following files
-------------------------------
*`/backend/perceptilabs/trainer/base.py`

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
I didn't make the codebase much better with this bug fix, but I took advantage of the good practices we've been employing (such as frequent refactors).

Related work items: #2069