Merged PR 1310: Fix deleted models persisting
# Implementation Summary
-------------------------------
This is a kind of intermittent issue.
I think the reason why it was happening is because the model data in IndexedDB is not deleted when we delete models.

# [Optional] How to review
-------------------------------
1. Create a model
2. Right click on a model in model list and select delete in the context menu
3. See Developer console > Application > IndexedDB > perceptilabs - idb > models
The proper data should be deleted.

# [Optional] Please review the following files
-------------------------------
N/A

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
We should avoid re-inventing the wheel.
We had the same logic for `delete_network` and `delete_networkbyid`, which had almost the same logic.
So I refactored `delete_network` to call `delete_networkbyid`

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2000