Merged PR 1342: [1826] Remove installations of gym['atari']
# Implementation Summary
-------------------------------
We no longer use the atari package, but we were instructing users to install it. It's been removed.
(There are still mentions of it in a template in the frontend, which I believe is being saved for something.)

# [Optional] How to review
-------------------------------
Make sure I didn't remove too much. Make sure the tool still runs.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
I removed extra references in the requirements file

# Please also:
~~- [ ] Include any new files in included_files.txt~~
- [X] Update the story on the board
~~- [ ] Announce changes that will affect everyone in #dev in slack~~

[1826] Remove installations of gym['atari']

Related work items: #1826