Merged PR 1344: Image Output Stats View
# Implementation Summary
-------------------------------
The goal of this story is to improve the statistics view for image outputs. Now, we have the predicted image in an overlay with the target image. This is the primary change for this story:

![Screen Shot 2021-06-17 at 2.57.46 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1344/attachments/Screen%20Shot%202021-06-17%20at%202.57.46%20PM.png)

# [Optional] How to review
-------------------------------
1. Start up the tool
2. Use an image to image dataset, like this one: [mnist_verysmall.zip](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1344/attachments/mnist_verysmall.zip)
3. Train, and you should see the target image in an overlay with the prediction

# [Optional] Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
I added a helper method at the top of `createDataObject.py` to assist in normalization for images. Our normalization was done in one line statements before, and in the context of the image processing done for this story,  it was done a second time unnecessarily. This was causing the image overlay to look differently from what it should've, which is why I added this method to help us normalize and de-normalize the data.

Related work items: #2008