Merged PR 1363: Added fixes for dot accuracy bug
# Implementation Summary
-------------------------------
To get rid of this bug, we return a data object with a singular point at the origin while the training statistics are being computed for the first epoch.

# [Optional] How to review
-------------------------------
Follow the repro steps in the bug ticket, and the empty chart should no longer appear.

# [Optional] Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
I tried to follow style guides for compound if statements and adhered to good general coding practices, such as not returning `None` outside of the conditionals added for this bug fix.

Related work items: #1959