Merged PR 1356: [2113] Add support in fileserver for parsing out BOM
# Implementation Summary
-------------------------------
We were not parsing files correctly when they had a BOM, including the BOM in the first line of the returned array of lines.
Instead, we now detect the encoding with the chardet module and use that to set the encoding.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Added a test using an file with a UTF-8 BOM

# Please also:
- [ ] Update the story on the board

[2113] Add support in fileserver for parsing out BOM

Related work items: #2113