Merged PR 1386: a image encoder recommendation ith 3 convolution layers
# Implementation Summary
-------------------------------
Earlier we only had a single conv layer in the recommended image encoder and it didn't work well with most datasets. After experimenting with few models and datasets, we're increasing the complexity of the image encoder with 3 conv layers that has shown promising performance on mnist, humanactivity, covid19 xray, blood cells, wildfires datasets.

# [Optional] How to review
-------------------------------
Load a dataset with image input and the recommended model should have 3 conv layers and a dense layer in the encoder

# [Optional] Please review the following files
-------------------------------
all

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
-

# Please also:
- [x] Include any new files in included_files.txt
- [x] Update the story on the board
- [x] Announce changes that will affect everyone in #dev in slack

Related work items: #2220