Merged PR 1388: [2210] Merge fileserver into rygg
# Implementation Summary
-------------------------------
To handle the datasets feature efficiently, we need fileserver and rygg running in the same process. With tricks for docker-compose volumes that we've added recently, the fileserver functionality doesn't need to run in the same container as the kernel.

# [Optional] How to review
-------------------------------
Make sure there's no remaining stuff from fileserver.
Make sure it runs for you.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
*Here you should describe what improvements you made to codebase while working on the ticket.*
*Note that the improvements should be proportional in size to the main objective of the ticket*

# Please also:
- [X] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

[2210] Merge fileserver into rygg

Related work items: #2210