Merged PR 1396: Cleans up 'checkCore' request from frontend and kernel
# Implementation Summary
-------------------------------
The purpose of this story was to remove the polling of the kernel's status every 2000 milliseconds and to remove non-specific "kernel is not connected" errors.

# How to review
-------------------------------
Start the kernel and create/train a model, make sure everything is still working on that front.

# Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
By removing unused code in both the kernel in the frontend, I hope that I've taken the right steps to de-spaghettify the parts of the codebase that need it the most.

Related work items: #2181