Merged PR 1397: String labels now show in the frontend
# Implementation Summary
-------------------------------
This PR changes the way training statistics for models built around categorical data are shown. Ideally, it'll make users' lives better by making categories more easily interpreted. Now, if the user has string categories for their data, it will show those strings in the frontend as well rather than just an integer as it did before. To do so, changes in the frontend and the kernel were made.

In the kernel, we get the string categories from the postprocessing pipeline and decode them if necessary (sometimes they're sent as bytes rather than strings). They're then sent to the frontend. Here's what it looks like now:

![Screen Shot 2021-07-13 at 2.45.18 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1397/attachments/Screen%20Shot%202021-07-13%20at%202.45.18%20PM.png)

# How to review
-------------------------------
1. Start up the tool
2. Try running a model with this dataset: [mnist_small_string.zip](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1397/attachments/mnist_small_string.zip)
3. Look at the training statistics, the x-axis of the chart representing the quantity predicted vs. actual category bar chart should have the categories now.

# Please review the following files
-------------------------------
**Backend**
* `/backend/perceptilabs/data/pipelines/categorical.py`
* `/backend/perceptilabs/layers/iooutput/stats/categorical.py`
* `/backend/perceptilabs/createDataObject.py`

**Frontend**
* `/frontend/src/renderer/core/mixins/net-element-viewBox.js`

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
This was a fairly small task, although it had a lot of moving parts. To make the improvements durable and code base a better place, I tried to make my changes cover all edge cases. For example, the parameters passed into `createDataObject()` are not always consistent, so I added logic to make sure that it can handle all possible combinations.

Related work items: #2228