Merged PR 1409: Adds Redis caching for pipeline metadata
# Implementation Summary
-------------------------------
This PR allows the kernels to share metadata for preprocessing. This gives a significant speedup when modeling and when starting training, since the metadata is only computed once.

Please note that this type of caching is not enabled per default in the pip/dev version of the tool, but it will be available for enterprise once #2261 is complete.

# [Optional] How to review
-------------------------------

1. spin up a redis Docker container: `docker run --network=host --name redis -e ALLOW_EMPTY_PASSWORD=yes bitnami/redis:latest`
2. Run the tool as usual, but with an extra environment variable: `PL_KERNEL_REDIS_PASSWORD=""` (should be visible to the Kernel and the Rendering Kernel).

(for convenience on Windows, you can skip the env var by modifying the `get_data_metadata_cache` function in `backend/perceptilabs/cache_utils.py`

What to expect:

* The initial data loading time will be a bit slower
* All subsequent calls will be a lot faster
* You should see printouts like this one: `2021-07-27 10:26:58,321 - INFO - perceptilabs.data.base:138 - Loaded preprocessing pipelines from metadata.` both when modifying the model setting and when starting training

# [Optional] Please review the following files
-------------------------------
N/A

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Factors out code for loading the data from cache. Each view reliant on data now reuses it.

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2248