Merged PR 1400: Update dataset settings on worksapce
# Implementation Summary
-------------------------------
*Here you can write about how the change affect the users and/or the developers.*
*You can also add details of **what** was changed, **why** it was changed and **how** it was changed.*

# [Optional] How to review
-------------------------------
1. Create a model
2. Go to workspace
3. Click Dataset button

You should be able to update dataset settings and pre processing options.
You also should be able to replace the dataset file with the one with the same fields.

Good example is brain tumor model and mnist_small model, and make sure that they have the same field names.
Once you change on to another, you will see the model broken because they have different number of neurons, but if you change it, the model should work

# [Optional] Please review the following files
-------------------------------
N/A

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Created a common module to remove duplicates and make it maintainable

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2127