Merged PR 1441: Single Env Variable for Redis
# Implementation Summary
-------------------------------
Before, we would not use the `REDIS_URL` as the reds password when deciding which cache to use between the `LightweightCache` and the `RedisCache`. This would default to the former for the metadata cache and the preview cache, and the goal of this task was to use the latter for everything instead.

# [Optional] How to review
-------------------------------
1. Start the tool up using redis and celery following @<Anton Kollmats> ' doc: https://mint-stem-e0c.notion.site/Enterprise-scaling-kernel-side-of-things-24601f83187b4272a1f1f00f53e2b261
2. Create MNIST
3. Train/Test/Export

# [Optional] Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Small changes in this PR, changed the name of `redis_password` to `redis_url` since we don't actually use a password for redis, and since the url is used more as a flag to determine whether or not we're using redis for the preview cache and metadata cache.

Related work items: #2275