Merged PR 1451: Models created base on datasets free/enterprise handling
# Implementation Summary
-------------------------------
This PR is based on #2123  feature it's include changes from #2210  and #2233

# [Optional] How to review
-------------------------------
In you want to run app in enterprise mode in In rygg/README.md is specified how to do this.
To run frontend properly for enterprise version in `constants.js` should be changed `ENTERPRISE_DATASET_FOLDER_PREFIX` with folder where was chosen to be stored uploaded files.

# [Optional] Please review the following files
-------------------------------
N/A

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
*Here you should describe what improvements you made to codebase while working on the ticket.*
*Note that the improvements should be proportional in size to the main objective of the ticket*

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #1588, #2210, #2233