Merged PR 1500:  Adds a single point of access for stats and checkpoints
# Implementation Summary
-------------------------------

Adds a single point of access for stats and checkpoints. File paths and so forth are encapsulated in an Epoch access object, where each epoch is accessed by ID.

The IDs used is the "training session ID", currently implemented using the checkpoint directory path. This will be replaced with a non-localized ID when we start working on model versioning.

The rationale for this change is to allow for upcoming changes related to cloud storage.

# [Optional] How to review
-------------------------------
Check that everything related to checkpoints works as usual:

* export
* model test
* resume from latest epoch
* serving [remember to enable feature flag]

# [Optional] Please review the following files
-------------------------------
N/A

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------

By encapsulating checkpoint access we can easily switch out the underlying storage without having to locate all files touching it.

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2362, #2363