Merged PR 1491: Pipeline build status
# Implementation Summary
-------------------------------
This story is for adding status updates to the pipelines, so that there is an update every time one is built. That update is a string which is pushed to Redis. We get the status update by passing in a callback function created in the PutData endpoint in `/backend/perceptilabs/endpoints/data/base.py` to the DataLoader module, which creates the pipelines through the `PipelineBuilder`.

In the `PutData` endpoint, the callback function `on_status_updated` is defined, which creates the status message and pushes it to Redis.

In the `IsDataReady` endpoint, the status is received from the cached metadata, in the event that the endpoint is called before the metadata has been cached, a default message is returned instead.

# [Optional] Please review the following files
-------------------------------
* `/backend/perceptilabs/caching/base.py`
* `/backend/perceptilabs/data/pipelines/base.py`
* `/backend/perceptilabs/endpoints/data/base.py`

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Wrote unit tests for the changes I made

Related work items: #2344