Merged PR 1532: stopping the training after end of iteration when stop request is sent
# Implementation Summary
-------------------------------
Training was earlier stopped only at the end of epoch when stop request is sent. Now, _loop_over_dataset method keep tracks of the stop variable and stops the training at the end of current iteration when stop request is sent.

# [Optional] How to review
-------------------------------
Start training. Stop the training. Check if the training has stopped in the kernel immediately or not.

# [Optional] Please review the following files
-------------------------------
all

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
-

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2299