Merged PR 1533: More granular status for loading message
# Implementation Summary
-------------------------------
Getting a more granular status update makes sense for computing preprocessing metadata, which iterates through the dataset and performs a datatype specific operation on each element of the dataset depending on which types of preprocessing are selected by the user. Certain types of preprocessing, especially for images, can take a while. The lack of visual feedback to the user on the build progress of their model recommendation is problematic, and this PR seeks to remedy that.

To fix it, I leveraged the callback function from the PR associated with #2344 and took it a step further by passing it from the base pipeline to the preprocessing pipelines that traverse the entire dataset. From there, the callback function passed in via a parameter to the preprocessing pipeline method `compute_metadata()` is invoked, and the current index and the size/cardinality of the dataset are used to construct a status message in the endpoint where the callback is originally defined.

# [Optional] How to review
-------------------------------
Create a model using the blood cell dataset and using normalization on the image feature. There should be a status message with a raw fraction showing the total samples processed so far out of total samples in the dataset.

# [Optional] Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Leveraged a wrapper function to prevent large numbers of changes to all of the pipelines. Rather than pass in all of the callback function's parameters as parameters to each of the pipelines, I defined a wrapper function (which calls the callback function) and pass it as the callback function to each of the pipelines.

Related work items: #2345