Merged PR 1544: Deleting latest-training-results.pkl file when training is started
# Implementation Summary
-------------------------------
Frontend fetches training results from the saved file immediately when training is started. this causes the frontend to show end results of previous training immediately when the model is trained for second time.

In order to fix this, now, kernel deletes the results.pkl file from previous run when a new CoreLogic is created.

# [Optional] How to review
-------------------------------
train any model for two times and see if the statistics are working as intended

# [Optional] Please review the following files
-------------------------------
all

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
-

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2388