Merged PR 1559: Using checkpoints and previous state to continue training
# Implementation Summary
-------------------------------
One of the issues happened to be, in Anton's refactoring of coreInterface, some code was left out that didn't let the trainer to use the previous state and checkpoints. Anton made a fix for it in hotfix/initial_stats_missing branch. This PR contains code from that branch as well.

Now, when the user uses checkpoints for training, we show stats from previous run as well during training to make it look like training is continuing.

A bug that would let the kernel delete the existing checkpoints incase of training failure has also been fixed.

# [Optional] How to review
-------------------------------
train a model. make sure the checkpoint files exist. Now restart the training and opt to train from latest epoch. Now, you should see the statistics from the previous training as well.

# [Optional] Please review the following files
-------------------------------
all

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
-
# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2458