Merged PR 1594: Show error messages when training or testing fails. Also fixes bug with kicking user back to workspace
# Implementation Summary
-------------------------------

This PR:

* adds error handling for training and testing jobs
* makes sure the user is kicked back to the workspace if training fails
* adds error handling for preview errors
* standardizes errors coming back from the HTTP API

There are now two types of failures shown to the user (aside from the lightweight core errors):

* caught errors: these are **expected errors**. For example, if the user has chosen an illegal setting we would raise such an error with an informative description. See the final screenshot below.
* uncaught errors: these are **unexpected errors**. These are true failures and are created automatically. They typically include the traceback.

The goal is to reduce the second type by eliminating bugs or by converting them to the first type so that the user gets concise and relevant information.

-------------------------------

For Kernel devs: the easiest way to propagate a specific/custom error message to the frontend is to raise a KernelError anywhere in your code (except in tasks and flask setup code).

For Frontend devs: the new convention is that the Kernel returns an "error" field for requests that raise an error. This structure contains a short "message" and additional info under "details" (typically the Kernel traceback).

# [Optional] How to review
-------------------------------

Start the tool and create settings that cause errors in the Kernel. If the Kernel turns out to be too robust, you can always raise exceptions in the code (see screenshots below for example).

## Examples:

Uncaught error during type inference:

![image (4).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1594/attachments/image%20%284%29.png)

Uncaught error during model recommendations:

![image (5).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1594/attachments/image%20%285%29.png)

Uncaught error during preview generation:

![image (6).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1594/attachments/image%20%286%29.png)

Uncaught error during training:
![image (3).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1594/attachments/image%20%283%29.png)

Uncaught error during testing:
![image (7).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1594/attachments/image%20%287%29.png)

Caught error during testing:
![image (2).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1594/at...