Merged PR 1558: Mixpanel tracking for pricing at the end of training (stopped/completed)
# Implementation Summary
-------------------------------
This pull request adds pricing info to mix panel tracking for when training stops/is completed. For more information on the parameters we use for pricing and how they're computed, see the following link:

#2460

# [Optional] Please review the following files
-------------------------------
* `/backend/perceptilabs/tracking/training_completed.py`
* `/backend/perceptilabs/tracking/training_stopped.py`
* `/backend/perceptilabs/trainer/base.py`

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
This pull request is small, so making the changes easy to read and leveraging the modularity we already have with mix panel tracking was a way to make these changes durable.

Related work items: #2460